feat: consolidate `system_time` functionality into `date.rs` and enhance `strftime` formatting with new specifiers and Ethiopian calendar support.
